% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roundtable.R
\name{roundtable_debug}
\alias{roundtable_debug}
\title{Debug a cluster run}
\usage{
roundtable_debug(key)
}
\arguments{
\item{key}{Key used when starting the tasks with \code{\link[=roundtable_run]{roundtable_run()}}}
}
\value{
A "task_bundle" object. Most likely you'll want to extract
individual tasks (\code{t <- grp$tasks[[1]]}) and look at their
errors (\code{t$result()} and \code{t$result()$trace}) and logs
(\code{t$log()}).
}
\description{
Debug a cluster run. Must be run on the cluster.
}
