% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roundtable.R
\name{roundtable_import}
\alias{roundtable_import}
\title{Import a roundtable job into orderly}
\usage{
roundtable_import(key)
}
\arguments{
\item{key}{The key created by \code{\link[=roundtable_rtm_prepare]{roundtable_rtm_prepare()}} and run
with \code{\link[=roundtable_run]{roundtable_run()}}}
}
\description{
Import a roundtable job back into orderly.
}
