% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtm.R
\name{roundtable_rtm_import}
\alias{roundtable_rtm_import}
\title{Import an RTM run into orderly}
\usage{
roundtable_rtm_import(key)
}
\arguments{
\item{key}{The key created by \code{\link[=roundtable_rtm_prepare]{roundtable_rtm_prepare()}} and run
with \code{\link[=roundtable_rtm_run]{roundtable_rtm_run()}}}
}
\description{
Import an RTM run back into orderly.
}
