% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtm.R
\name{roundtable_rtm_prepare}
\alias{roundtable_rtm_prepare}
\title{Prepare a rtm run for use on the cluster.}
\usage{
roundtable_rtm_prepare(
  name,
  test_run = FALSE,
  short_run = FALSE,
  date = NULL,
  chains = 16L
)
}
\arguments{
\item{name}{The name of the task (see \code{\link[orderly:orderly_list]{orderly::orderly_list()}})}

\item{test_run}{Run a short test run on the server before export,
to validate that the data and code play well together.}

\item{short_run}{On the cluster, should this be a short run? Set
this to \code{TRUE} only for debugging.}

\item{date}{The date to run}

\item{chains}{The number of chains to run. Note that this
duplicates the default in the underlying task so be aware of
updating either of these numbers.}
}
\value{
The key
}
\description{
Prepare a rtm run for use on the cluster. This should be run on
the main server, from master. Once run, note the key, which you
will want for use with \code{\link[=roundtable_run]{roundtable_run()}} on the cluster.
}
