% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtm.R
\name{roundtable_rtm_run}
\alias{roundtable_rtm_run}
\title{Run RTM job on cluster}
\usage{
roundtable_rtm_run(key, test_job = FALSE, upgrade = FALSE, rerun = FALSE)
}
\arguments{
\item{key}{The job key}

\item{test_job}{Send a test job to the cluster to ensure that
everything works.}
}
\value{

}
\description{
Run a prepared RTM job on the cluster. Requires a key as created
by \code{\link[=roundtable_rtm_prepare]{roundtable_rtm_prepare()}}. Run this on the cluster, from a
directory that is \emph{not} the checked out copy of the orderly
sources, as this will not interact with it.
}
\author{
Richard Fitzjohn
}
