% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roundtable.R
\name{roundtable_status}
\alias{roundtable_status}
\title{Summary status}
\usage{
roundtable_status(key)
}
\arguments{
\item{key}{A key, as created by \code{\link[=roundtable_rtm_prepare]{roundtable_rtm_prepare()}}}
}
\value{
Primarily called for its side effect, though some data is
returned in a list.
}
\description{
Print summary status for a single key
}
